Pare down register options
After I finally got this man page slimmed down, I (re-?)realized
the purpose of this command to install packages directly from
ports, which is probably why it was eliminated before.

However, all the kyua tests are using this.  It's probably the
same code necessary for packing.  I've got to look at this further
to see if we keep it or eliminate it again.

I think we might need to keep it for testsuite purposes.