suite is all wrong
The word "pkg" is so overused, the mass-replace really jacked things
up.  I'm not sure how "new_pkg" got changed to "new_manifest" but
this whole thing is a mess.  Every test will have to be reimported
and hand modified.

It doesn't help that extension was changed to ".pkg" for freebsd's
pkg program.  ughh.